+++ src/main/java/pe/edu/upc/ariseapp/servicesinterfaces/IRolService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface IRolService {
     public List<Rol> list();
     public void insert(Rol ro);
-    public Rol listId(int idRol);
+    public Rol listId(Long idRol);
     public void update(Rol ro);
-    public void delete(int idRol);
+    public void delete(Long idRol);
 }
